requirements.txt
# Core Framework
streamlit>=1.38.0
langchain==1.0.0
langchain-community==0.4
langchain-text-splitters==1.0.0
langchain-core==1.0.0
langgraph==1.0.0
python-dotenv>=1.0.0

# LLM Providers
langchain-nvidia-ai-endpoints>=0.1.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-ollama>=0.2.0

# RAG & Embeddings
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy>=1.24.0,<2.0
torch==2.4.1  # CPU-only by default; install CUDA build separately if needed
rank-bm25==0.2.2

# Document Processing & OCR
PyMuPDF==1.24.9  # fitz
pytesseract==0.3.13
Pillow==10.4.0  # For OCR images
beautifulsoup4==4.12.3
lxml==5.3.0  # Parser for BS4

# HTTP & API
requests>=2.31.0
raganything>=1.0.0

# Database & Utils
# sqlite3 is built-in to Python (no pip package)
# If migrating to SQLAlchemy: uncomment the following
# sqlalchemy==2.0.32

# Dev Tools (optional, for VS Code)
black==24.8.0
isort==5.13.2
pytest==8.3.2